requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
